Strip comments, prettify html and fix multiple configurations bug
Multiple configuration would have used the same file to compare old vs new
which was prevented by adding a unique name attribute to the configuration.
Also to run on a raspberry pi where the default python3 is 3.6 explicitly 3.8 is
now specified.
Also some " were converted to ' if no " was necessary.